+++ app/javascript/mastodon/features/follow_requests/components/account_authorize.js
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 import Permalink from '../../../components/permalink';
+import AccountRelationshipButtonContainer from '../../../containers/account_relationship_button_container';
 import Avatar from '../../../components/avatar';
 import DisplayName from '../../../components/display_name';
 import IconButton from '../../../components/icon_button';
@@ -23,9 +24,38 @@ export default class AccountAuthorize extends ImmutablePureComponent {
     intl: PropTypes.object.isRequired,
   };
 
+  state = {
+    requestState: 'requested',
+  };
+
+  onClickAuthorize = () => {
+    this.props.onAuthorize();
+    this.setState({
+      requestState: 'authorized',
+    });
+  }
+
+  onClickReject = () => {
+    this.props.onReject();
+    this.setState({
+      requestState: 'rejected',
+    });
+  }
+
   render () {
-    const { intl, account, onAuthorize, onReject } = this.props;
+    const { intl, account } = this.props;
+    const { requestState } = this.state;
     const content = { __html: account.get('note_emojified') };
+
+    const accountPanel = () => {
+      const disabled = requestState !== 'requested';
+      return (
+        <div className='account--panel'>
+          <IconButton className={`account--panel__button ${requestState === 'authorized' ? requestState : ''}`} title={intl.formatMessage(messages.authorize)} icon='check' onClick={this.onClickAuthorize} disabled={disabled} />
+          <IconButton className={`account--panel__button ${requestState === 'rejected' ? requestState : ''}`} title={intl.formatMessage(messages.reject)} icon='times' onClick={this.onClickReject} disabled={disabled} />
+        </div>
+      );
+    };
 
     return (
       <div className='account-authorize__wrapper'>
@@ -35,13 +65,13 @@ export default class AccountAuthorize extends ImmutablePureComponent {
             <DisplayName account={account} />
           </Permalink>
 
+          <div className='account-authorize__relationship'>
+            <AccountRelationshipButtonContainer id={account.get('id')} size={22} />
+          </div>
           <div className='account__header__content' dangerouslySetInnerHTML={content} />
         </div>
 
-        <div className='account--panel'>
-          <div className='account--panel__button'><IconButton title={intl.formatMessage(messages.authorize)} icon='check' onClick={onAuthorize} /></div>
-          <div className='account--panel__button'><IconButton title={intl.formatMessage(messages.reject)} icon='times' onClick={onReject} /></div>
-        </div>
+        {accountPanel()}
       </div>
     );
   }
